refactor(Server): add result to handle_post_request
This commit refactors the handle_post_request function to include a result parameter, which is used to store the result of the POST request.

The previous implementation of handle_post_request did not return any value, which made it difficult to use the function in situations where the result of the POST request was needed. This change addresses that issue by adding a result parameter to the function, which can be used to store the result of the POST request.

This change improves the usability of the handle_post_request function by making it possible to use the function in situations where the result of the POST request is needed.